requirements.txt
Cython
setuptools==57.0.0
moviepy==1.0.3
Pillow==8.2.0
PyYAML==6.0.1
typer==0.12.3
